+++ CHANGELOG.md
@@ -8,7 +8,7 @@ CHANGELOG
 - Add `mem_used_percent` and `disk_used_percent` metrics for head node memory and root volume disk utilization tracking on the ParallelCluster CloudWatch dashboard, and set up alarms for monitoring these metrics.
 - Add log rotation support for ParallelCluster managed logs.
 - Track common errors of compute nodes on Cloudwatch Dashboard.
-- Increase the limit on the maximum number of queues per cluster from 10 to 40.
+- Increase the limit on the maximum number of queues per cluster from 10 to 100. Each cluster can however have a maximum number of 150 compute resources and each queue can have a maximum of 40 compute resources.
 - Allow to specify a sequence of multiple custom actions scripts per event.
 - Add support for customizing the cluster Slurm configuration via the ParallelCluster configuration YAML file.
 - Track the longest dynamic node idle time in CloudWatch Dashboard.